Fixed some cases don't execute cleanCache() function

+++ build.sh
@@ -2,7 +2,7 @@
 #
 #Flyme auto make script
 
-buildDir=/home/leon/Android/Flyme/build
+buildDir=/home/leon/android/flyme/build
 pwd=$PWD
 
 function initEnv() {
@@ -24,28 +24,27 @@ function makeFull() {
 		
         mv flyme*.zip Flyme_A0001_Leon_$versionName
         mv target*.zip target_files.zip
-		
+
+        cd ..	
         echo "<<< 全量包编译完成"
     else
         echo "<<< 全量包编译失败"
     fi
-
-    cd ..
 }
 
 function makeOTA() {
     echo "是否编译OTA包？(Y/n)"
 		
     read keyboard
+
     case $keyboard in
         Y|y|YES|yes)
-        echo ">>> 编译OTA包"
-		
-        ../../build/tools/releasetools/ota_from_target_files -k ../../build/security/testkey -i history_package/last_target_files.zip out/target_files.zip out/OTA_Flyme_A0001_$versionName
-        cleanCache
-
+        echo ">>> 编译OTA包"	
+        $buildDir/tools/releasetools/ota_from_target_files -k $buildDir/security/testkey -i history_package/last_target_files.zip out/target_files.zip out/OTA_Flyme_A0001_$versionName
         echo "<<< OTA包编译完成"
-    esac	
+    esac
+
+    cleanCache	
 }
 
 function cleanCache(){
@@ -58,7 +57,10 @@ function cleanCache(){
 
         mv out/target_files.zip history_package
         mv out/Flyme*.zip Flyme_Done
-        mv out/OTA*.zip Flyme_Done
+
+        if [ -e out/OTA*.zip ];then
+            mv out/OTA*.zip Flyme_Done
+        fi
 
         mv history_package/target_files.zip history_package/last_target_files.zip
     fi
